requirements.txt
asyncio
PyYAML
telethon
python-dotenv
